Add lib code for better handling of implications
Added Implicator class for applying implications to tag sets, but only regular
implications. This is to ensure that implication will not get caught in a
circular relation and that an implication will not be obscured by an alias.
It also places a restriction on aliases: a tag alias cannot be part of an
implication.

Added new methods to TagSet and refactored existing methods for handling
implications of any type.