Enable warnings on improper TagActions use
Warn in two situations: if a set/union name is over-written by later
assignment, or if a non-existent set/union name is referenced by a chain/union
assignment. It remains the case that the order of defining unions matters.
With a defaultdict the empty set was produced on a missing key, which in a
chain caused the whole chain to fail. Now at least the user is warned about
this occurrence to better protect from typos.