Exclude empty values from semicolon list
This changes split_semicolon_list so that empty or whitespace-only values in
the list are not returned in the final list.
Before, a config value of ';;' would return a list of three empty strings.
Now, excess semicolons, such as trailing semicolons, don't add extraneous,
empty values to the result.
I can't conceive of a *current* use case for lists with empty strings in one's
configuration.