Change Implicator.implications to read-only
Create a new iterator method 'all_implications' that exposes implications in
the form of RegularImplications and make the attribute 'implications' a
read-only, compatibility form of that. This way Implicator does not have to
maintain two forms of the same data: 'graph' from ImplicationGraph, a
defaultdict[str, TagSet], and its own 'implications', a
set[RegularImplication]. The graph alone is sufficient.